refactor(app): move homepage internals to a private folder
Every directory under `app/` claims a URL segment. `app/home/` held 19
files of homepage internals and no `page.tsx`, so `/home` fell through
to the not-found page -- but only by the absence of one file, and
nothing in the name said the folder was not meant to be a route.

Next.js excludes a folder whose name starts with `_` from routing
entirely, so `app/_home/` cannot serve `/home` however many files land
in it. The invariant is now a property of the name rather than of what
nobody has added yet.

Mechanical: 19 files moved with `git mv`, 11 import specifiers updated
across three files, `tsc --noEmit` clean, no `app/home` reference left
anywhere including docs. No route added or removed and no behaviour
changed.

The new browser assertion is a characterisation test, not a red/green
one: `/home` returned the not-found page before this change too, so it
passes on either side. It is worth keeping because it pins the
acceptance criterion directly rather than trusting that nobody adds a
`page.tsx`, and the existing unknown-route journey checks a different
path.

AGENTS.md records the convention, so the next internal folder under
`app/` follows it.

Closes #142

Entire-Checkpoint: 01M04CHY42S6ZS4639MREQ6D01
