+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This project follows a lightweight, human-curated changelog. Keep the newest cha
 
 ### Changed
 
+- The homepage's internal components and modules moved from `app/home/` to `app/_home/`. Every directory under `app/` claims a URL segment, so a folder holding internals was one file away from becoming a public route at `/home`, with nothing in its name to say it should not be. A leading underscore marks it private to Next.js, which excludes it from routing outright. No route was added or removed and nothing about the app changed for a visitor; `/home` returns the branded not-found page exactly as before, and now cannot stop doing so.
 - The app no longer downloads a web font, and a production build no longer needs Google Fonts to be reachable. The root layout fetched Geist and Geist Mono at build time, but the Tailwind font tokens have always defined system stacks and nothing ever referenced the `--font-geist-*` variables the layout generated, so no text was rendered in Geist — a visitor's browser only preloaded two font files it never used. Rendered typography is unchanged, the system stack it actually uses is documented beside the tokens that define it, and the build now depends on nothing outside the repository for fonts.
 - Label synchronization now decides what to do in tested code. `scripts/sync-labels.mjs` had grown to mix validating `.github/labels.yml`, planning creates and updates, and deleting labels, with only the offline file check covered in CI — so the destructive branch was the least tested part of it. Validation, planning, reporting, and applying now live in `scripts/label-plan.mjs` as pure functions over plain data, unit tested without a token or a network: that a deletion is planned only with `--prune`, that a dry run previews exactly the deletions a real run would make and writes nothing, that an unquoted colour like `123456` is caught while `e10c02` passes through as the string it stays, and that what reaches the GitHub API is what the plan validated. The script itself is unchanged in behaviour and still the only entry point; `npm run check:labels` still validates the file offline on every pull request.
 - The unit test run now collects from `app/`, `components/`, and `scripts/` instead of skipping directories one at a time. A scratch spec left at the repository root, or in any directory the exclude list had not been taught about, used to join `npm test` and fail the gate with a Playwright error that pointed at the tooling rather than at the stray file. Nothing outside those three roots is a unit test now, so such a file neither breaks the run nor quietly joins it, and the verification exemption for the external tooling-state directories no longer depends on a denylist keeping pace with every directory a tool might create. `npm run check:agent-docs` still fails if the ignore lists drift apart, and now also if the collection scope widens.
