+++ app/PrivacySafeAnalytics.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Analytics } from "@vercel/analytics/next";
-import { redactAnalyticsEvent } from "./home/analytics";
+import { redactAnalyticsEvent } from "./_home/analytics";
 
 export default function PrivacySafeAnalytics() {
   return <Analytics beforeSend={redactAnalyticsEvent} />;
