+++ app/_home/SearchErrorState.tsx
