+++ app/_home/SearchShortcutSection.tsx
