+++ app/_home/analytics.ts
